refactor(core): Decouple post workflow execute event from internal hooks (no-changelog)
Follow-up to: #10221
